Update Composer's configuration
This change adds in packages required to run quality assurance checks
and testing of the code. It also adds an autoloader and dev autoloader
configuration as the code will be refactored from the anonymous function
into a class shortly. Then, it adds in some scripts to help with running
tests and QA checks.